Python-Vet (english version).
This is a program that builds a clinical history of a patient brought to consultation. It gathers the basic data of the patient, such as name, age and species; asked to the owner.
Then it helps you get to a diagnostic -for that it asks you for the temperature, weight, pulse, cardiac rate, respiratory rate and capillary refill. For ending our clinical history, it shows a physical examination of the pet that determines the body condition it has due to the veterinary criteria.