Backend_development_just_for_functionality
In this repository I'll upload the first and simplest code made for backend of apps before applying a framework. This is backend that focuses only on the functionality of the program and doesn't have GUI for the web or desktop.

This consists in a group of ideas of how the backend is expected to do its workflow, and it could be used in the future for the development of an app for web, desktop or mobile.