Alarm Clock.
This is a program that allows you to set an alarm clock in a period of time chosen by the user. You can use a custom audio file for this.

NOTE: The use of this module (winsound) is set only for Windows. You may need to import other modules for other OS.